// src/app/taf-fiscal/tsi/models/tsi-divergent-documents-item.ts
export interface TsiDivergentDocumentsItem {
  branchcode: string;
  notefoundinsft: string;
  notefoundintaf: string;
  integrationerrorinv5r: string;
  typingdate: string;
  operationtype: string;
  documentnumber: string;
  series: string;
  participantcodeandname: string;
  participantscnpj: string;
  cancelederp: string;
  canceledtaf: string;
  stampc20: string;
  stampsft: string;
  participantcode: string;
  store: string;
  recnov5r: number;
  details: string;
}